+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/validators/ProcessValidator.java
@@ -50,21 +50,21 @@ public class ProcessValidator implements Validator<Process> {
 
   private void validateForCreateAndUpdate(Process process) throws ValidationException {
     requireExistingDomain(process);
-    process.getZones().forEach(zone -> {
+    process.getZoneKeys().forEach(zone -> {
       if (!zoneView.exists(zone)) {
         throw new ValidationException("Zone [" + zone + "] does not exist");
       }
     });
 
     process.getInputs().forEach(input -> {
-      if (!streamView.exists(input.getStream())) {
-        throw new ValidationException("Input stream [" + input.getStream() + "] does not exist");
+      if (!streamView.exists(input.getStreamKey())) {
+        throw new ValidationException("Input stream [" + input.getStreamKey() + "] does not exist");
       }
     });
 
     process.getOutputs().forEach(output -> {
-      if (!streamView.exists(output.getStream())) {
-        throw new ValidationException("Output stream [" + output.getStream() + "] does not exist");
+      if (!streamView.exists(output.getStreamKey())) {
+        throw new ValidationException("Output stream [" + output.getStreamKey() + "] does not exist");
       }
     });
   }