+++ core/src/test/java/com/expediagroup/streamplatform/streamregistry/core/services/ProcessServiceTest.java
@@ -327,7 +327,7 @@ public class ProcessServiceTest {
     p.setKey(new ProcessKey("domain","name"));
     List<ZoneKey> zones = new ArrayList<>();
     zones.add(new ZoneKey("aws_us_east_1"));
-    p.setZones(zones);
+    p.setZoneKeys(zones);
     return p;
   }
 }