+++ core/src/test/java/com/expediagroup/streamplatform/streamregistry/core/validators/ProcessValidatorTest.java
@@ -15,6 +15,9 @@
  */
 package com.expediagroup.streamplatform.streamregistry.core.validators;
 
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
@@ -29,9 +32,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.junit.Before;
 import org.junit.Test;
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
@@ -193,7 +193,7 @@ public class ProcessValidatorTest {
     p.setKey(new ProcessKey("domain","name"));
     List<ZoneKey> zones = new ArrayList<>();
     zones.add(new ZoneKey("aws_us_east_1"));
-    p.setZones(zones);
+    p.setZoneKeys(zones);
     return p;
   }
 }