+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Changed
+- producer & consumer creation not allowed when domain is not available
+- Fixed `streamKey` field name in `ProcessInputStream` and `ProcessOutputStream`
+- Fixed `streamKey` field name in `ProcessInputStreamInput` and `ProcessOutputStreamInput`
+- Fixed `zoneKeys` field in `Process`
+- Fixed `zoneKey` field in `ProcessBinding`
+
 ## [1.1.2] 2022-08-05
 ### Updated
 - Process creation not allowed when domain is not available